Move PoolTitle and TokenIcon to top level
